test: performance test to get_str added
